+++ graphql/resolvers/booking.js
@@ -1,0 +1,50 @@
+'use strict';
+
+const Booking = require('../../models/booking');
+const { dateToString } = require('../../helpers/date');
+const { singleEvent, user } = require('./util');
+
+
+const transformBooking = booking => {
+  return {
+    ...booking._doc,
+    _id: booking.id,
+    user: user.bind(this, booking._doc.user),
+    event: singleEvent.bind(this, booking._doc.event),
+    createdAt: dateToString(booking._doc.createdAt),
+    updatedAt: dateToString(booking._doc.updatedAt)
+  };
+};
+
+
+module.exports = {
+  bookings: async () => {
+    try {
+      const bookings = await Booking.find();
+      return bookings.map(booking => {
+        return transformBooking(booking);
+      });
+    } catch(err){
+      throw err;
+    }
+  },
+  bookEvent: async (args) => {
+    const fetchedEvent = await Event.findOne({_id: args.eventId});
+    const booking = new Booking({
+      user: '5d8641449ee9a92fbc53ef75',
+      event: fetchedEvent
+    });
+    const result = await booking.save();
+    return transformBooking(result);
+  },
+  cancelBooking: async args => {
+    try{
+      const booking = await Booking.findById(args.bookingId).populate('event');
+      const event = transformEvent(booking.event);
+      await Booking.deleteOne({_id: args.bookingId});
+      return event;
+    }catch(err){
+      throw err;
+    }
+  }
+};
