+++ graphql/resolvers/index.js
@@ -1,154 +1,13 @@
 'use strict';
 
-const bcrypt = require('bcryptjs');
+const authResolver = require('./auth');
+const eventsResolver = require('./events');
+const bookingResolver = require('./booking');
 
-const Event = require('../../models/event');
-const User = require('../../models/user');
-const Booking = require('../../models/booking');
-const { dateToString } = require('../../helpers/date');
-
-const transformEvent = event => {
-  return {
-    ...event._doc,
-    _id: event.id,
-    date: dateToString(event._doc.date),
-    creator: user.bind(this, event.creator)
-  };
-}
-
-const events = async eventIds => {
-  try {
-    const events = await Event.find({_id: {$in: eventIds} });
-    return events.map(event => {
-      return transformEvent(event);
-    });
-
-  }catch(err) {
-      throw err;
-    }
+const rootResolver = {
+  ...authResolver,
+  ...eventsResolver,
+  ...bookingResolver
 };
 
-const singleEvent = async eventId => {
-  try {
-    const event = await Event.findById(eventId);
-    return transformEvent(event);
-  }catch(err){
-    throw err;
-  }
-}
-
-const user = async (userId) => {
-  try {
-    const user = await User.findById(userId);
-    return {
-      ...user._doc,
-      _id: user.id,
-      createdEvents: events.bind(this, user._doc.createdEvents)
-    };
-  }catch(err){
-    throw err;
-  }
-}
-
-module.exports = {
-  events: async () => {
-    try {
-      const events = await Event.find();
-      return events.map(event => {
-        console.log('events:', event._doc);
-        return transformEvent(event);
-      });
-    }catch(err){
-      throw err; // express will catch this err and pass down to error handler
-    }
-  },
-  bookings: async () => {
-    try {
-      const bookings = await Booking.find();
-      return bookings.map(booking => {
-        return {
-          ...booking._doc,
-          _id: booking.id,
-          user: user.bind(this, booking._doc.user),
-          event: singleEvent.bind(this, booking._doc.event),
-          createdAt: dateToString(booking._doc.createdAt),
-          updatedAt: dateToString(booking._doc.updatedAt)
-        };
-      });
-    } catch(err){
-      throw err;
-    }
-  },
-  createEvent: async (args) => {
-    const event = new Event({
-      title: args.eventInput.title,
-      description: args.eventInput.description,
-      price: +args.eventInput.price,
-      date: dateToString(args.eventInput.date),
-      creator: '5d8641449ee9a92fbc53ef75'
-    });
-    let createdEvent;
-    try {
-      const result = await event.save();
-      console.log('createEvent:', result._doc);
-      createdEvent = transformEvent(result);
-      console.log('event.save(): ', result);
-      const creator = await User.findById('5d8641449ee9a92fbc53ef75');
-      if(!creator){
-        throw new Error('user not found');
-      }
-      creator.createdEvents.push(event);
-      await creator.save();
-      return createdEvent;
-    }catch(err){
-      console.log(err);
-      throw err;
-    }
-  },
-  createUser: async (args) => {
-    try {
-      let existingCreator = await User.findOne({email: args.userInput.email});
-      if(existingCreator){
-        throw new Error('user exists already.');
-      }
-      const hashedPassword = await bcrypt.hash(args.userInput.password, 12);
-      const creator = new User({
-        email: args.userInput.email,
-        password: hashedPassword
-      });
-      const userSaveResult = await creator.save();
-      return {...userSaveResult._doc, password: null, _id: userSaveResult.id};
-    }catch(err){
-      throw err;
-    }
-  },
-
-  bookEvent: async (args) => {
-    const fetchedEvent = await Event.findOne({_id: args.eventId});
-    const booking = new Booking({
-      user: '5d8641449ee9a92fbc53ef75',
-      event: fetchedEvent
-    });
-    const result = await booking.save();
-    return {
-      ...result._doc,
-      _id: result.id,
-      user: user.bind(this, booking._doc.user),
-      event: singleEvent.bind(this, booking._doc.event),
-      createdAt: dateToString(result._doc.createdAt),
-      updatedAt: dateToString(result._doc.updatedAt)
-    };
-  },
-
-  cancelBooking: async args => {
-    try{
-      const booking = await Booking.findById(args.bookingId).populate('event');
-      const event = transformEvent(booking.event);
-      await Booking.deleteOne({_id: args.bookingId});
-      return event;
-    }catch(err){
-      throw err;
-    }
-
-  }
-};
+module.exports = rootResolver;
