+++ graphql/resolvers/util.js
@@ -1,0 +1,39 @@
+'use strict';
+
+const events = async eventIds => {
+  try {
+    const events = await Event.find({_id: {$in: eventIds} });
+    return events.map(event => {
+      return transformEvent(event);
+    });
+
+  }catch(err) {
+      throw err;
+    }
+};
+
+const singleEvent = async eventId => {
+  try {
+    const event = await Event.findById(eventId);
+    return transformEvent(event);
+  }catch(err){
+    throw err;
+  }
+}
+
+const user = async (userId) => {
+  try {
+    const user = await User.findById(userId);
+    return {
+      ...user._doc,
+      _id: user.id,
+      createdEvents: events.bind(this, user._doc.createdEvents)
+    };
+  }catch(err){
+    throw err;
+  }
+}
+
+exports.events = events;
+exports.singleEvent = singleEvent;
+exports.user = user;
